# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Dashboard App API",
    "description": "Dashboard App API",
    "termsOfService": "Terms of usage v1",
    "contact": {
      "name": "Sangram Nandkhile",
      "email": "sangram@example.com"
    }
  },
  "paths": {
...
    "/api/Users/{Id}/password": {
      "patch": {
        "tags": [
          "User Management"
        ],
        "summary": "Update the password",
        "operationId": "UpdatePassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "description": "UserId",
            "required": true,
            "type": "string",
            "format": "string"
          },
          {
            "name": "Password",
            "in": "body",
            "description": "Password Dto",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Password"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          },
          "404": {
            "description": "Supplied reference was not found",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          },
          "500": {
            "description": "Internal server Error",
            "schema": {
              "$ref": "#/definitions/Exception"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Password": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    },
    "Exception": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization:Bearer{token}\" "
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}